+++ code/core/rttbDVHSet.cpp
@@ -209,7 +209,7 @@ namespace rttb
 		{
 			s << "[ ";
 
-			for (unsigned int i = 0; i < aDVHSet.size(); i++)
+			for (size_t i = 0; i < aDVHSet.size(); i++)
 			{
 				s << aDVHSet.at(i);
 			}
